Self-Checkout Systems
Project Scope: Making a self checkout (cash register without cashier) system with Raspberry pi, checking the products with a weight sensor and making payments with local pay integration.

About the Project : 

This project is a self-checkout system with a time limit of 2 months. 
HTML, Boostrap 5 and JavaScript are used for front-end in the program. 
For the back-end side, API's were created with Python language and Firebase database system integration was made.